Enable tests by default, and add "reset" option to autogen
OASIS upstream justify tests being disabled by default by saying that
"dependencies of `Test` will be excluded by default". However, this is an
unsafe approach, and it is always better to default to safety. Furthermore,
the desire to "reduce the number of dependencies" is (with a smooth enough
package manager) quite an esoteric requirement, and doesn't need to be
catered for by default.